# openapi.yaml
openapi: 3.0.3
info:
  title: Professional Pride Room Booking API
  version: 2.0.0
  description: |
    Comprehensive REST API for room booking management system with advanced features including:
    - Room availability validation and conflict prevention
    - Customer management with ID proof verification
    - Dynamic pricing based on room configurations
    - Excel-style booking grid with horizontal scrolling
    - Payment processing with screenshot uploads
    - Invoice generation and PDF export
    - Real-time dashboard analytics
    - Mobile-optimized responsive interfaces
    
    **Key Features:**
    - Prevents double bookings with real-time validation
    - Supports multiple ID proof uploads per customer
    - Dynamic pricing based on number of guests
    - Comprehensive booking history tracking
    - Professional invoice generation
    - Mobile-responsive design
servers:
  - url: https://your-service.onrender.com
    description: Production server
  - url: http://localhost:8082
    description: Local development server
paths:
...
  /api/v1/auth/login:
    post:
      summary: Staff login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [userId, password]
              properties:
                userId: 
                  type: string
                  description: Staff user ID
                password: 
                  type: string
                  description: Staff password
      responses:
        '200': 
          description: Login successful
          content: 
            application/json: 
              schema: 
                type: object
                properties:
                  token: 
                    type: string
                  userType: 
                    type: string
                    enum: [staff]
  /api/v1/auth/logout:
    post:
      summary: Staff logout
      parameters:
        - in: header
          name: Authorization
          required: false
          schema: 
            type: string
      responses:
        '200': 
          description: Logout successful
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
security:
  - BearerAuth: []